App.py divided, data folder added
