+++ cleanup.sh
